Update many optimizations
Optimize `addBatches` to call methods on the underlying
`PreparedStatement` directly instead of using `set` and `addBatch`.

A new method `updateManyReturningGeneratedKeys` has been created which
works the same as `updateManyWithGeneratedKeys` but returns a collection
instead of a stream. Use in cases where a known value of rows were given
and the overhead of having a stream is not required.