Ensure no connections older than max lifetime are returned
Currently when a connection is returned it is checked to ensure it is younger than the maximum
configured connection lifetime. However only the first connection retrieved from the pool is checked,
the fallback is not. So check the age of the fallback to ensure no connections are given out that
exceed the maximum confirued connection lifetime.