Ensure no connections older than max lifetime are returned
Currently when a connection is returned it is checked to ensure it is
younger than the maximum configured connection lifetime. However only
the first connection retrieved from the pool is checked, the fall-back
is not. So check the age of the fall-back to ensure no connections are
given out that exceed the maximum configured connection lifetime.